+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 - Button to clear address field
 
+### Changed:
+- Improments in the Watch-Only login view
+
 ## [2.0.1] - 2024-03-05
 ### Added:
 - Manual coin selection for L-BTC